+++ src/enclave/main.cpp
@@ -36,7 +36,8 @@ namespace ccf
     size_t num_worker_threads,
     const ccf::ds::WorkBeaconPtr& work_beacon,
     ccf::AbstractRuntimeControl& runtime_control,
-    const std::shared_ptr<AbstractLedgerSubsystemInterface>& ledger_subsystem)
+    const std::shared_ptr<AbstractLedgerSubsystemInterface>& ledger_subsystem,
+    const std::shared_ptr<AbstractNodeTransport>& node_transport)
   {
     std::lock_guard<ccf::ds::Mutex> guard(create_lock);
 
@@ -51,8 +52,13 @@ namespace ccf
       return CreateNodeStatus::EnclaveInitFailed;
     }
 
-    // Setup logger to allow enclave logs to reach the host before node is
-    // actually created
+    if (node_transport == nullptr)
+    {
+      LOG_FAIL_FMT("A node transport must be provided to create a node");
+      return CreateNodeStatus::EnclaveInitFailed;
+    }
+
+    // The host-enclave circuit, from which the enclave reads host messages
     auto circuit = std::make_unique<ringbuffer::Circuit>(
       ringbuffer::BufferDef{
         enclave_config.to_enclave_buffer_start,
@@ -62,10 +68,6 @@ namespace ccf
         enclave_config.from_enclave_buffer_start,
         enclave_config.from_enclave_buffer_size,
         enclave_config.from_enclave_buffer_offsets});
-    auto basic_writer_factory =
-      std::make_unique<ringbuffer::WriterFactory>(*circuit);
-    auto writer_factory = std::make_unique<oversized::WriterFactory>(
-      *basic_writer_factory, enclave_config.writer_config);
 
     {
       num_pending_threads = (uint16_t)num_worker_threads + 1;
@@ -109,8 +111,6 @@ namespace ccf
       // NOLINTBEGIN(cppcoreguidelines-owning-memory)
       enclave = new ccf::Enclave(
         std::move(circuit),
-        std::move(basic_writer_factory),
-        std::move(writer_factory),
         ccf_config.ledger_signatures.tx_count,
         ccf_config.ledger_signatures.delay.count_ms(),
         enclave_config.tick_interval,
@@ -120,7 +120,8 @@ namespace ccf
         ccf_config.node_certificate.curve_id,
         work_beacon,
         runtime_control,
-        ledger_subsystem);
+        ledger_subsystem,
+        node_transport);
       // NOLINTEND(cppcoreguidelines-owning-memory)
     }
     catch (const std::exception& exc)